# api/docs/swagger.yaml
definitions:
  handler.Response:
    properties:
      data: {}
      description:
        type: string
      status:
        type: integer
    type: object
...
  models.CreateOqituvchi:
    properties:
      FanTarmogi:
        type: string
      date_of_birth:
        type: string
      email:
        type: string
      first_name:
        type: string
      ish_joyi:
        type: string
      last_name:
        type: string
      mavzular:
        type: string
      middle_name:
        type: string
      mutahassislik:
        type: string
      password:
        type: string
      phone_number:
        type: string
    type: object
...
  models.UpdateOqituvchi:
    properties:
      FanTarmogi:
        type: string
      date_of_birth:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      ish_joyi:
        type: string
      last_name:
        type: string
      mavzular:
        type: string
      middle_name:
        type: string
      mutahassislik:
        type: string
      password:
        type: string
      phone_number:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /oqituvchi:
    get:
      consumes:
      - application/json
      description: Get List Oqituvchi
      operationId: get_list_oqituvchi
      parameters:
      - description: offset
        in: query
        name: offset
        type: string
      - description: limit
        in: query
        name: limit
        type: string
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get List Oqituvchi
      tags:
      - Oqituvchi
    post:
      consumes:
      - application/json
      description: Create Oqituvchi
      operationId: create_oqituvchi
      parameters:
      - description: CreateOqituvchiRequest
        in: body
        name: oqituvchi
        required: true
        schema:
          $ref: '#/definitions/models.CreateOqituvchi'
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Create Oqituvchi
      tags:
      - Oqituvchi
  /oqituvchi/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Oqituvchi
      operationId: delete_oqituvchi
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Delete Oqituvchi
      tags:
      - Oqituvchi
    get:
      consumes:
      - application/json
      description: Get By ID Oqituvchi
      operationId: get_by_id_oqituvchi
      parameters:
      - description: id
        in: path
        name: id
        type: string
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get By ID Oqituvchi
      tags:
      - Oqituvchi
    put:
      consumes:
      - application/json
      description: Update Oqituvchi
      operationId: update_oqituvchi
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: UpdateOqituvchiRequest
        in: body
        name: oqituvchi
        required: true
        schema:
          $ref: '#/definitions/models.UpdateOqituvchi'
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update Oqituvchi
      tags:
      - Oqituvchi
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"